% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PMID.R
\name{get_PMID}
\alias{get_PMID}
\title{Choose today's PMID}
\usage{
get_PMID(subject = "genomics")
}
\arguments{
\item{subject}{a string for lotto category}
}
\value{
PMID for today
}
\description{
Choose today's PMID
}
\examples{
get_PMID()
get_PMID("neurobiology")
}
