% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery.R
\name{lottery}
\alias{lottery}
\title{Play the PMID lottery}
\usage{
lottery(lotto_numbers, subject = "genomics")
}
\arguments{
\item{lotto_numbers}{a integer vector of PMID guesses}

\item{subject}{a string for lotto category}
}
\value{
Whether or not numbers match today's PMID, and information on the paper
}
\description{
Play the PMID lottery
}
\examples{
lottery(38863995)
lottery(c(38863995, 38864247, 38862712, 38862494))

}
