% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery_hint.R
\name{lottery_hint}
\alias{lottery_hint}
\title{Hint for today's numbers}
\usage{
lottery_hint(subject = "genomics")
}
\arguments{
\item{subject}{a string for lotto category}
}
\value{
A hint for today's possible PMID range
}
\description{
Hint for today's numbers
}
\examples{
lottery_hint()
}
