% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery.R
\name{lottery}
\alias{lottery}
\title{Title}
\usage{
lottery(lotto_numbers)
}
\arguments{
\item{lotto_numbers}{a integer vector of PMID guesses}
}
\value{
Any guess matches today's PMID
}
\description{
Title
}
\examples{
lottery(38863995)
lottery(c(38863995, 38864247, 38862712, 38862494))

}
